automerge commit
git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/astum@71149 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/rtp.c
@@ -1050,7 +1050,7 @@ struct ast_frame *ast_rtcp_read(struct ast_rtp *rtp)
 								    "IAJitter: %u\r\n"
 								    "LastSR: %lu.%010lu\r\n"
 								    "DLSR: %4.4f(sec)\r\n"
-								    "RTT: %lu(sec)\r\n",
+								    "RTT: %llu(sec)\r\n",
 								    ast_inet_ntoa(sin.sin_addr), ntohs(sin.sin_port),
 								    pt, (pt == 200) ? "Sender Report" : (pt == 201) ? "Receiver Report" : (pt == 192) ? "H.261 FUR" : "Unknown",
 								    rc,
@@ -1062,7 +1062,7 @@ struct ast_frame *ast_rtcp_read(struct ast_rtp *rtp)
 								    rtp->rtcp->reported_jitter,
 								    (unsigned long) ntohl(rtcpheader[i + 4]) >> 16, ((unsigned long) ntohl(rtcpheader[i + 4]) << 16) * 4096,
 								    ntohl(rtcpheader[i + 5])/65536.0,
-								    rtt);
+								    (unsigned long long)rtt);
 			} else {
 				manager_event(EVENT_FLAG_REPORTING, "RTCPReceived", "From %s:%d\r\n"
 								    "PT: %d(%s)\r\n"
